Moves remaining components from ManageAuthorPage to its own file
