+++ frontend/src/Features/ManageEntity/Author/Parts/ManageAuthorHeader.tsx
@@ -1,0 +1,11 @@
+import { Divider } from '@chakra-ui/react';
+
+export default function ManageAuthorHeader() {
+  return (
+    <>
+      <h1 className="font-inter mb-2 text-4xl font-medium text-black">Manage Author</h1>
+      <h2 className="font-roboto mb-6 text-lg text-gray-500">You can see available authors and create new author here.</h2>
+      <Divider mb="1.5rem" />
+    </>
+  );
+}
